# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovedRevisionResponse": {
        "additionalProperties": false,
        "description": "One immutable ApprovedRevision and its re-derived Ready qualification.\n\n`ready_qualified` is a property of this revision's own stored evidence.\nWhether it is the *active* Ready milestone is `preparation_state` on\n`GET /applications/{id}`, and the two are separate on purpose: a revision\nstays qualified and downloadable after a newer JobSnapshot has moved the\nApplication off it.",
        "properties": {
          "application_id": {
            "title": "Application Id",
            "type": "string"
          },
          "approved_at": {
            "title": "Approved At",
            "type": "string"
          },
          "decision_provenance": {
            "additionalProperties": true,
            "title": "Decision Provenance",
            "type": "object"
          },
          "draft_content_hash": {
            "title": "Draft Content Hash",
            "type": "string"
          },
          "draft_edit_version": {
            "title": "Draft Edit Version",
            "type": "integer"
          },
          "facts_version": {
            "title": "Facts Version",
            "type": "string"
          },
          "html_artifact_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Html Artifact Version Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job_analysis_id": {
            "title": "Job Analysis Id",
            "type": "string"
          },
          "job_snapshot_id": {
            "title": "Job Snapshot Id",
            "type": "string"
          },
          "pdf_artifact_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Artifact Version Id"
          },
          "ready_qualified": {
            "title": "Ready Qualified",
            "type": "boolean"
          },
          "ready_validation": {
            "$ref": "#/components/schemas/ValidationReportResponse"
          },
          "selection_plan_id": {
            "title": "Selection Plan Id",
            "type": "string"
          },
          "validation_run_id": {
            "title": "Validation Run Id",
            "type": "string"
          },
          "version_number": {
            "title": "Version Number",
            "type": "integer"
          },
          "working_draft_id": {
            "title": "Working Draft Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "application_id",
          "version_number",
          "working_draft_id",
          "job_snapshot_id",
          "job_analysis_id",
          "selection_plan_id",
          "validation_run_id",
          "draft_edit_version",
          "draft_content_hash",
          "facts_version",
          "approved_at",
          "decision_provenance",
          "ready_qualified",
          "ready_validation"
        ],
        "title": "ApprovedRevisionResponse",
        "type": "object"
      },
...
      "DecisionMarkdownResponse": {
        "additionalProperties": false,
        "description": "One revision's provenance as a human-readable document.\n\nThe suggested save name travels in `Content-Disposition`, the way the\nrecruiter PDF's does, rather than as a body field: a download name is a\ntransport concern, and a schema property shaped like a filename is\nindistinguishable from a stored location to anything reading the contract.\n\n`content_hash` names exactly what was produced, so a saved copy stays\ncheckable against the record it came from.",
        "properties": {
          "application_id": {
            "title": "Application Id",
            "type": "string"
          },
          "approved_revision_id": {
            "title": "Approved Revision Id",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "content_hash": {
            "title": "Content Hash",
            "type": "string"
          }
        },
        "required": [
          "application_id",
          "approved_revision_id",
          "content",
          "content_hash"
        ],
        "title": "DecisionMarkdownResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperationAction": {
        "enum": [
          "cancel",
          "retry"
        ],
        "title": "OperationAction",
        "type": "string"
      },
      "OperationFailureCode": {
        "enum": [
          "SOURCE_CHANGED",
          "PROVIDER_TIMEOUT",
          "PROVIDER_RATE_LIMITED",
          "PROVIDER_UNAVAILABLE",
          "PROVIDER_REFUSED",
          "INVALID_OUTPUT",
          "SCHEMA_VIOLATION",
          "RENDER_FAILED",
          "BROWSER_START_FAILED",
          "VALIDATION_EXECUTION_FAILED",
          "CANCELLED_BEFORE_ACTIVATION"
        ],
        "title": "OperationFailureCode",
        "type": "string"
      },
...
      "OperationPhase": {
        "enum": [
          "queued",
          "waiting_for_application",
          "waiting_for_render_slot",
          "waiting_for_ai_slot",
          "pre_execution_check",
          "executing",
          "retry_wait",
          "pre_activation_check",
          "activating",
          "completed"
        ],
        "title": "OperationPhase",
        "type": "string"
      },
      "OperationResponse": {
        "additionalProperties": false,
        "description": "The §11 Operation query fields, and nothing wider.\n\nThis mirrors `OperationView`, not `PersistedOperation`. The payload, the\nfrozen sources, the lease, and the idempotency key stay inside the runner.\n\n`message` is a safe progress line and `safe_failure_detail` a safe failure\nline; neither carries a path, a provider response, or a key. There is no\npercentage field, because the specification forbids fabricating one.\n\nThe closed sets are typed as the application enums rather than flattened to\n`str`, so the generated TypeScript is a real union instead of `string`, and\n`is_terminal` is reported rather than left to the client. Which statuses end\nan Operation is a lifecycle rule this layer already owns; a client that\nre-derives it keeps a second copy of that rule, and the copy is what goes\nstale when the lifecycle gains a status.\n\n`is_terminal` and `available_actions` are computed from lifecycle state rather\nthan accepted as fields.\nThis representation is built from two places - `operation_response` and the\n`active_operation` of an application projection - and a field would have to\nbe supplied correctly at both. A computed value has no second place to\nforget.",
        "properties": {
          "application_id": {
            "title": "Application Id",
            "type": "string"
          },
          "available_actions": {
            "description": "Commands currently accepted, derived by the application layer.",
            "items": {
              "$ref": "#/components/schemas/OperationAction"
            },
            "readOnly": true,
            "title": "Available Actions",
            "type": "array"
          },
          "cancellation_requested_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancellation Requested At"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "failure_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OperationFailureCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_terminal": {
            "description": "Derived from the same predicate the runner uses.",
            "readOnly": true,
            "title": "Is Terminal",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "operation_type": {
            "$ref": "#/components/schemas/OperationType"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/OperationOutputResponse"
            },
            "title": "Outputs",
            "type": "array"
          },
          "phase": {
            "$ref": "#/components/schemas/OperationPhase"
          },
          "retry_of_operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Of Operation Id"
          },
          "safe_failure_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safe Failure Detail"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/OperationStatus"
          }
        },
        "required": [
          "id",
          "application_id",
          "operation_type",
          "status",
          "phase",
          "message",
          "created_at",
          "outputs",
          "is_terminal",
          "available_actions"
        ],
        "title": "OperationResponse",
        "type": "object"
      },
      "OperationStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "cancelled",
          "interrupted"
        ],
        "title": "OperationStatus",
        "type": "string"
      },
      "OperationType": {
        "enum": [
          "analyze_job",
          "propose_selection_plan",
          "create_draft",
          "regenerate_section",
          "regenerate_claim",
          "render_revision"
        ],
        "title": "OperationType",
        "type": "string"
      },
...
      "RenderRevisionRequest": {
        "additionalProperties": false,
        "description": "The Application the client believes it is rendering for.\n\nExplicit rather than inferred from the revision, for the same reason\n`apply-decisions` states it: a client that names both is telling the server\nwhat it believes, and a mismatch is a `412` naming the broken lineage\ninstead of a render landing on another Application's revision.",
        "properties": {
          "application_id": {
            "title": "Application Id",
            "type": "string"
          }
        },
        "required": [
          "application_id"
        ],
        "title": "RenderRevisionRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "ValidationReportResponse": {
        "additionalProperties": false,
        "properties": {
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "groups": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Groups",
            "type": "object"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssueResponse"
            },
            "title": "Issues",
            "type": "array"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          },
          "report_schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Schema Version"
          }
        },
        "required": [
          "passed",
          "groups",
          "issues",
          "evidence"
        ],
        "title": "ValidationReportResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CV Engine local API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/approved-revisions/{approved_revision_id}": {
      "get": {
        "description": "`200` with the immutable record and a freshly re-derived qualification (§20).\n\nA revision superseded by a newer JobSnapshot still answers here, still\nreports `ready_qualified`, and is still exportable. What it stops being is\nthe Application's active `preparation_state`, which is a different question\nasked at a different resource.",
        "operationId": "approved_revision_detail_api_v1_approved_revisions__approved_revision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "approved_revision_id",
            "required": true,
            "schema": {
              "title": "Approved Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovedRevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read one approved revision and its Ready qualification",
        "tags": [
          "approved-revisions"
        ]
      }
    },
    "/api/v1/approved-revisions/{approved_revision_id}/decision-markdown": {
      "get": {
        "description": "`200` and the document; `409` when the revision belongs elsewhere.\n\nReturned as content rather than streamed as a file. This is the\nhuman-readable provenance record - what was selected, which gaps were\naccepted, which overrides were recorded - and a client that receives it as\ntext can show it beside the revision it explains. The suggested save name\ntravels in `Content-Disposition`, and `content_hash` names exactly what was\nproduced, so a caller that saves it can prove later which export it holds.",
        "operationId": "export_decision_markdown_api_v1_approved_revisions__approved_revision_id__decision_markdown_get",
        "parameters": [
          {
            "in": "path",
            "name": "approved_revision_id",
            "required": true,
            "schema": {
              "title": "Approved Revision Id",
              "type": "string"
            }
          },
          {
            "description": "The Application this revision is expected to belong to. Stated by the caller rather than inferred, so a mismatch is a refusal naming the broken lineage instead of an export of another Application's provenance.",
            "in": "query",
            "name": "application_id",
            "required": true,
            "schema": {
              "description": "The Application this revision is expected to belong to. Stated by the caller rather than inferred, so a mismatch is a refusal naming the broken lineage instead of an export of another Application's provenance.",
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionMarkdownResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export human-readable provenance for one approved revision",
        "tags": [
          "approved-revisions"
        ]
      }
    },
    "/api/v1/approved-revisions/{approved_revision_id}/preview": {
      "get": {
        "operationId": "preview_approved_revision_api_v1_approved_revisions__approved_revision_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "approved_revision_id",
            "required": true,
            "schema": {
              "title": "Approved Revision Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "html_artifact_version_id",
            "required": true,
            "schema": {
              "title": "Html Artifact Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The exact approved HTML, safe to frame in an isolated iframe."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview one exact verified rendered HTML artifact",
        "tags": [
          "approved-revisions"
        ]
      }
    },
    "/api/v1/approved-revisions/{approved_revision_id}/recruiter-pdf": {
      "get": {
        "description": "`200` and the PDF; `412` when this exact pair is not Ready-qualified.\n\nBoth IDs are checked against each other before qualification is computed,\nso a PDF belonging to another revision is `412 LINEAGE_BROKEN` rather than\na qualification run against a mismatched pair.",
        "operationId": "export_recruiter_pdf_api_v1_approved_revisions__approved_revision_id__recruiter_pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "approved_revision_id",
            "required": true,
            "schema": {
              "title": "Approved Revision Id",
              "type": "string"
            }
          },
          {
            "description": "The exact rendered PDF to export. Required and explicit: an export that resolved the latest PDF for itself could hand over a different document than the one the caller verified.",
            "in": "query",
            "name": "pdf_artifact_version_id",
            "required": true,
            "schema": {
              "description": "The exact rendered PDF to export. Required and explicit: an export that resolved the latest PDF for itself could hand over a different document than the one the caller verified.",
              "title": "Pdf Artifact Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "The rendered PDF, named as a recruiter should see it."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export the exact Ready PDF under its recruiter-facing filename",
        "tags": [
          "approved-revisions"
        ]
      }
    },
    "/api/v1/approved-revisions/{approved_revision_id}/render": {
      "post": {
        "description": "`202` and a `Location`: rendering is a durable Operation (§16).\n\nA failed render leaves the ApprovedRevision approved. Nothing in the render\npath writes to the revision - it is immutable - so the failure is recorded\non the Operation and in a `rendered-invalid` artifact lifecycle, and the\nrevision is exactly as approvable-from as it was. Retrying is\n`POST /operations/{id}/retry`, which creates a *new* Operation rather than\nreopening the failed one.\n\nThe same `Idempotency-Key` with the same payload returns the Operation it\nalready created instead of queueing a second render.",
        "operationId": "render_revision_api_v1_approved_revisions__approved_revision_id__render_post",
        "parameters": [
          {
            "in": "path",
            "name": "approved_revision_id",
            "required": true,
            "schema": {
              "title": "Approved Revision Id",
              "type": "string"
            }
          },
          {
            "description": "Optional. A retry that reuses the key of an Operation which already exists returns that Operation instead of queueing a second attempt. Omitted, the boundary generates a key, exactly as the CLI does.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional. A retry that reuses the key of an Operation which already exists returns that Operation instead of queueing a second attempt. Omitted, the boundary generates a key, exactly as the CLI does.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderRevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Render one exact approved revision",
        "tags": [
          "approved-revisions"
        ]
      }
    },
...
    "/api/v1/artifacts/{artifact_version_id}/download": {
      "get": {
        "description": "`200` and the bytes; `412` when the stored payload does not verify.\n\nA `404` means no such registration. A `412` means the registration exists\nand its payload failed containment, presence, or its hash - three separate\ncodes, because \"somebody moved it\" and \"somebody changed it\" are different\nfindings and a client should not have to guess which it hit.\n\n`Content-Length` is set from the size the store measured after verifying\nthe hash, so a client can show progress against a number that was true of\nthe exact bytes being sent.",
        "operationId": "download_artifact_api_v1_artifacts__artifact_version_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_version_id",
            "required": true,
            "schema": {
              "title": "Artifact Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "The registered payload, under a safe delivery filename."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download one registered artifact by ID",
        "tags": [
          "artifacts"
        ]
      }
    },
...
    "/api/v1/operations/{operation_id}/retry": {
      "post": {
        "description": "`202` and a `Location` naming the *new* Operation.\n\nThe original is immutable and is not touched. Only a terminal Operation can\nbe retried; retrying a live one is a `409`.",
        "operationId": "retry_operation_api_v1_operations__operation_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "title": "Operation Id",
              "type": "string"
            }
          },
          {
            "description": "Optional. A retry that reuses the key of an Operation which already exists returns that Operation instead of queueing a second attempt. Omitted, the boundary generates a key, exactly as the CLI does.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional. A retry that reuses the key of an Operation which already exists returns that Operation instead of queueing a second attempt. Omitted, the boundary generates a key, exactly as the CLI does.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue a new Operation retrying a terminal one",
        "tags": [
          "operations"
        ]
      }
    },
...
  }
}